refactor(service): 纯移动拆分 openai_ws_forwarder.go（4675→399行）
